remove the video resource query in favor of adapter
