+++ apps/course-builder-web/src/app/(content)/tips/[slug]/edit/page.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { notFound } from 'next/navigation'
 import { Layout } from '@/components/app/layout'
+import { courseBuilderAdapter } from '@/db'
 import { getTip } from '@/lib/tips-query'
-import { getVideoResource } from '@/lib/video-resource-query'
 import { getServerAuthSession } from '@/server/auth'
 
 import { EditTipForm } from '../../_components/edit-tip-form'
@@ -23,7 +23,7 @@ export default async function TipEditPage({
 
 	const resource = tip.resources?.[0]?.resource.id
 
-	const videoResourceLoader = getVideoResource(resource)
+	const videoResourceLoader = courseBuilderAdapter.getVideoResource(resource)
 
 	return (
 		<Layout>
