+++ apps/course-builder-web/src/app/(content)/tutorials/[module]/[lesson]/page.tsx
@@ -5,11 +5,11 @@ import { headers } from 'next/headers'
 import Link from 'next/link'
 import { notFound } from 'next/navigation'
 import { TipPlayer } from '@/app/(content)/tips/_components/tip-player'
+import { courseBuilderAdapter } from '@/db'
 import { Lesson } from '@/lib/lessons'
 import { getLesson } from '@/lib/lessons-query'
 import { Tutorial } from '@/lib/tutorial'
 import { getTutorial } from '@/lib/tutorials-query'
-import { getVideoResource } from '@/lib/video-resource-query'
 import { getServerAuthSession } from '@/server/auth'
 import { getOGImageUrlForResource } from '@/utils/get-og-image-url-for-resource'
 import ReactMarkdown from 'react-markdown'
@@ -189,7 +189,7 @@ async function PlayerContainer({
 
 	const resource = lesson.resources?.[0]?.resource.id
 
-	const videoResourceLoader = getVideoResource(resource)
+	const videoResourceLoader = courseBuilderAdapter.getVideoResource(resource)
 
 	return (
 		<Suspense fallback={<PlayerContainerSkeleton />}>
