+++ apps/pro-aws/src/app/(content)/tips/_components/new-resource-with-video-form.tsx
@@ -3,9 +3,9 @@
 import * as React from 'react'
 import { useRouter } from 'next/navigation'
 import { TipUploader } from '@/app/(content)/tips/_components/tip-uploader'
+import { courseBuilderAdapter } from '@/db'
 import { NewTip, NewTipSchema } from '@/lib/tips'
 import { createTip } from '@/lib/tips-query'
-import { getVideoResource } from '@/lib/video-resource-query'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useForm } from 'react-hook-form'
 
@@ -48,7 +48,8 @@ export function NewResourceWithVideoForm() {
 		let delay = initialDelay
 
 		for (let attempt = 1; attempt <= maxAttempts; attempt++) {
-			const videoResource = await getVideoResource(videoResourceId)
+			const videoResource =
+				await courseBuilderAdapter.getVideoResource(videoResourceId)
 			if (videoResource) {
 				yield videoResource
 				return
