+++ apps/pro-aws/src/lib/tips-query.ts
@@ -9,7 +9,6 @@ import {
 	contributionTypes,
 } from '@/db/schema'
 import { Tip, TipSchema, type NewTip, type TipUpdate } from '@/lib/tips'
-import { getVideoResource } from '@/lib/video-resource-query'
 import { getServerAuthSession } from '@/server/auth'
 import { guid } from '@/utils/guid'
 import slugify from '@sindresorhus/slugify'
@@ -100,7 +99,9 @@ export async function createTip(input: NewTip) {
 
 	const newTipId = `tip_${guid()}`
 
-	const videoResource = await getVideoResource(input.videoResourceId)
+	const videoResource = await courseBuilderAdapter.getVideoResource(
+		input.videoResourceId,
+	)
 
 	if (!videoResource) {
 		throw new Error('🚨 Video Resource not found')
